Move copilot related-files to language features
It is now in its own file.

Also address other PR comments.
